added delete confirmation js popup window

+++ public/js/script.js
@@ -2,6 +2,13 @@
  * Created by Nika on 7/11/2015.
  */
 $(document).ready(function(){
+    var deleteEvent = function(){
+        $('.delete').click(function(e){
+            if(!confirm('დარწმუნებული ხართ თუ არა რომ გსურთ ჩანაწერის წაშლა ?')){
+                e.preventDefault();
+            }
+        });
+    }
     $('#region').change(function() {
         var region   = $(this).val();
 
@@ -113,7 +120,7 @@ $(document).ready(function(){
             if(this.isAdmin){
                 div.append(' <div id = '+ id +'>' +
                 '<h3>' + name + '</h3>' +
-                '<a href="/delete/'+id+'"><span>წაშალე ტრენინგი</span></a>'+
+                '<a class="delete" onClick ="confirm("Press a button!")"  href="/delete/'+id+'"><span>წაშალე ტრენინგი</span></a>'+
                 '<div class="description">'+description+'</div>' +
                 '<div class="row"><div class="col-sm-3"><strong>სწავლების სფერო: </strong></div><div class="col-sm-9">'+field+'</div></div>'+
                 '<div class="row"><div class="col-sm-3"><strong>სწავლების ფორმა: </strong></div><div class="col-sm-9">'+terms+'</div></div>'+
@@ -135,6 +142,7 @@ $(document).ready(function(){
                 '</div><hr class="margin-bottom-40">');
             }
             });
+            deleteEvent();
         }).fail(function(){
             var div = $('#search-result');
             div.html('');
@@ -167,7 +175,7 @@ $(document).ready(function(){
                 if(this.isAdmin){
                     div.append(' <div id = '+ id +'>' +
                     '<h3>დასახელება: ' + name + '</h3>' +
-                    '<a href="/delete_seek/'+id+'"><span>წაშალე ტრენინგი</span></a>'+
+                    '<a class="delete"  href="/delete_seek/'+id+'">წაშალე ტრენინგი</a>'+
                     '<div class="description">'+description+'</div>' +
                     '<div class="row"><div class="col-sm-3"><strong>სწავლების სფერო: </strong></div><div class="col-sm-9">'+field+'</div></div>'+
                     '<div class="row"><div class="col-sm-3"><strong>მონაწილეთა რაოდენობა: </strong></div><div class="col-sm-9">'+quantity+'</div></div>'+
@@ -189,6 +197,7 @@ $(document).ready(function(){
                     '</div><hr class="margin-bottom-40">');
                 }
             });
+            deleteEvent();
         }).fail(function(){
             var div = $('#search-result');
             div.html('');
@@ -222,7 +231,7 @@ $(document).ready(function(){
                 if(this.isAdmin){
                     div.append(' <div id = '+ id +'>' +
                     '<h3>დასახელება: ' + name + '</h3>' +
-                    '<a href="/delete_seek/'+id+'"><span>წაშალე ტრენინგი</span></a>'+
+                    '<a class="delete" onClick ="confirm("Press a button!")"  href="/delete_seek/'+id+'"><span>წაშალე ტრენინგი</span></a>'+
                     '<div class="description">'+description+'</div>' +
                     '<div class="row"><div class="col-sm-3"><strong>სწავლების სფერო: </strong></div><div class="col-sm-9">'+field+'</div></div>'+
                     '<div class="row"><div class="col-sm-3"><strong>მონაწილეთა რაოდენობა: </strong></div><div class="col-sm-9">'+quantity+'</div></div>'+
@@ -245,6 +254,7 @@ $(document).ready(function(){
                 }
 
             });
+            deleteEvent();
         }).fail(function(){
             var div = $('#search-result');
             div.html('');
@@ -278,7 +288,7 @@ $(document).ready(function(){
                 if(this.isAdmin){
                     div.append(' <div id = '+ id +'>' +
                     '<h3>' + name + '</h3>' +
-                    '<a href="/delete/'+id+'"><span>წაშალე ტრენინგი</span></a>'+
+                    '<a class="delete" onClick ="confirm("Press a button!")"  href="/delete/'+id+'"><span>წაშალე ტრენინგი</span></a>'+
                     '<div class="description">'+description+'</div>' +
                     '<div class="row"><div class="col-sm-3"><strong>სწავლების სფერო: </strong></div><div class="col-sm-9">'+field+'</div></div>'+
                     '<div class="row"><div class="col-sm-3"><strong>სწავლების ფორმა: </strong></div><div class="col-sm-9">'+terms+'</div></div>'+
@@ -300,6 +310,7 @@ $(document).ready(function(){
                     '</div><hr class="margin-bottom-40">');
                 }
             });
+            deleteEvent();
         }).fail(function(){
             var div = $('#search-result');
             div.html('');
@@ -307,4 +318,6 @@ $(document).ready(function(){
         });
         event.preventDefault();
     });
+
+    
 });
